Move SolverConf out of verify
